Safari-611.1.10/LayoutTests/media/video-poster-set-after-playback-expected.txt


EVENT(playing)
EXPECTED (internals.elementShouldDisplayPosterImage(video) == 'false') OK
EVENT(pause)
EXPECTED (internals.elementShouldDisplayPosterImage(video) == 'false') OK
END OF TEST
